Replace sqlite3 with Postgres when deploying to Heroku.
I'm still not sure quite where to put the database backend abstraction
boundary. I'd prefer not to go all the way over to Django-style ORM,
because that feels like a lot of boilerplate; but I'm already having
to jump through hoops to get database cursors that work the same way
in sqlite3 and psycopg2.